Bug #2767 more float to double, projection working OK now

+++ Code/Gui/MITK/Modules/niftkOpenCV/VideoTrackerMatching/mitkVideoTrackerMatching.cxx
@@ -336,7 +336,7 @@ void VideoTrackerMatching::SetVideoLagMilliseconds ( unsigned long long VideoLag
 //---------------------------------------------------------------------------
 cv::Mat VideoTrackerMatching::ReadTrackerMatrix(std::string filename)
 {
-  cv::Mat TrackerMatrix = cv::Mat(4,4, CV_32FC1);
+  cv::Mat TrackerMatrix = cv::Mat(4,4, CV_64FC1);
   std::ifstream fin(filename.c_str());
   if ( !fin )
   {
@@ -347,7 +347,7 @@ cv::Mat VideoTrackerMatching::ReadTrackerMatrix(std::string filename)
   {
     for ( int col = 0 ; col < 4 ; col ++ ) 
     {
-      fin >> TrackerMatrix.at<float>(row,col);
+      fin >> TrackerMatrix.at<double>(row,col);
     }
   }
   return TrackerMatrix;
@@ -433,7 +433,7 @@ void VideoTrackerMatching::SetCameraToTracker (cv::Mat matrix, int trackerIndex)
 //---------------------------------------------------------------------------
 cv::Mat VideoTrackerMatching::GetTrackerMatrix ( unsigned int FrameNumber , long long * TimingError  ,unsigned int TrackerIndex  )
 {
-  cv::Mat returnMat = cv::Mat(4,4,CV_32FC1);
+  cv::Mat returnMat = cv::Mat(4,4,CV_64FC1);
   
   if ( !m_Ready ) 
   {
